backend-example/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Face Recognition Libraries
deepface==0.0.81
tf-keras==2.15.0
tensorflow==2.15.0

# Database (choose one)
# SQLAlchemy==2.0.23  # For SQL databases
# motor==3.3.2        # For MongoDB

# Optional: for production deployment
gunicorn==21.2.0
redis==5.0.1 